Updating bridge protocol, moving CapacitorBridge, adding deprecated stub for CAPBridge.
