+++ ios/Capacitor/Capacitor/JSExport.swift
@@ -15,32 +15,32 @@ internal class JSExport {
     public static func exportCapacitorJS(userContentController: WKUserContentController) throws {
         guard let jsUrl = Bundle.main.url(forResource: "public/native-bridge", withExtension: "js") else {
             CAPLog.print("ERROR: Required native-bridge.js file in Capacitor not found. Bridge will not function!")
-            throw BridgeError.errorExportingCoreJS
+            throw CapacitorBridgeError.errorExportingCoreJS
         }
 
         do {
             try self.injectFile(fileURL: jsUrl, userContentController: userContentController)
         } catch {
             CAPLog.print("ERROR: Unable to read required native-bridge.js file from the Capacitor framework. Bridge will not function!")
-            throw BridgeError.errorExportingCoreJS
+            throw CapacitorBridgeError.errorExportingCoreJS
         }
     }
 
     public static func exportCordovaJS(userContentController: WKUserContentController) throws {
         guard let cordovaUrl = Bundle.main.url(forResource: "public/cordova", withExtension: "js") else {
             CAPLog.print("ERROR: Required cordova.js file not found. Cordova plugins will not function!")
-            throw BridgeError.errorExportingCoreJS
+            throw CapacitorBridgeError.errorExportingCoreJS
         }
         guard let cordovaPluginsUrl = Bundle.main.url(forResource: "public/cordova_plugins", withExtension: "js") else {
             CAPLog.print("ERROR: Required cordova_plugins.js file not found. Cordova plugins  will not function!")
-            throw BridgeError.errorExportingCoreJS
+            throw CapacitorBridgeError.errorExportingCoreJS
         }
         do {
             try self.injectFile(fileURL: cordovaUrl, userContentController: userContentController)
             try self.injectFile(fileURL: cordovaPluginsUrl, userContentController: userContentController)
         } catch {
             CAPLog.print("ERROR: Unable to read required cordova files. Cordova plugins will not function!")
-            throw BridgeError.errorExportingCoreJS
+            throw CapacitorBridgeError.errorExportingCoreJS
         }
 
     }
